+++ Source/Shared/Storage/Entry.swift
