+++ Source/Shared/Storage/ImageWrapper.swift
