+++ Source/Shared/Storage/PrimitiveWrapper.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct PrimitiveWrapper<T: Codable> {
+  let value: T
+}
