+++ Source/Shared/StorageLegacy/CacheEntry.swift
