+++ Tests/iOS/Tests/Storage/ImageWrapperTests.swift
@@ -1,0 +1,9 @@
+//
+//  ImageWrapperTests.swift
+//  Cache-iOS-Tests
+//
+//  Created by Khoa Pham on 12.09.2017.
+//  Copyright © 2017 Hyper Interaktiv AS. All rights reserved.
+//
+
+import Foundation
