+++ Tests/iOS/Tests/Storage/PrimitiveWrapperTests.swift
@@ -1,0 +1,9 @@
+//
+//  PrimitiveWrapperTests.swift
+//  Cache-iOS-Tests
+//
+//  Created by Khoa Pham on 12.09.2017.
+//  Copyright © 2017 Hyper Interaktiv AS. All rights reserved.
+//
+
+import Foundation
